+++ animation_nodes/nodes/number/float_to_text.py
@@ -1,19 +1,35 @@
 import bpy
-from ... base_types import AnimationNode
+from ... base_types import AnimationNode, VectorizedSocket
 
 class FloatToTextNode(bpy.types.Node, AnimationNode):
     bl_idname = "an_FloatToTextNode"
     bl_label = "Float to Text"
 
+    useList: VectorizedSocket.newProperty()
+
     def create(self):
-        self.newInput("Float", "Number", "number")
+        self.newInput(VectorizedSocket("Float", "useList",
+            ("Number", "number"), ("Numbers", "numbers")))
         self.newInput("Integer", "Min Length", "minLength", value = 10, minValue = 0)
         self.newInput("Integer", "Decimals", "decimals", value = 3, minValue = 0)
         self.newInput("Boolean", "Insert Sign", "insertSign").value = False
-        self.newOutput("Text", "Text", "text")
 
-    def execute(self, number, minLength, decimals, insertSign):
-        sign = "+" if insertSign else ""
+        self.newOutput(VectorizedSocket("Text", "useList",
+            ("Text", "text"), ("Texts", "texts")))
 
-        formatString = "{" + ":{}0{}.{}f".format(sign, max(minLength, 0), max(decimals, 0)) + "}"
+    def getExecutionFunctionName(self):
+        if self.useList:
+            return "executeList"
+        return "executeSingle"
+
+    def executeSingle(self, number, minLength, decimals, insertSign):
+        formatString = self.getFormatString(minLength, decimals, insertSign)
         return formatString.format(number)
+
+    def executeList(self, numbers, minLength, decimals, insertSign):
+        formatString = self.getFormatString(minLength, decimals, insertSign)
+        return [formatString.format(number) for number in numbers]
+
+    def getFormatString(self, minLength, decimals, insertSign):
+        sign = "+" if insertSign else ""
+        return "{" + ":{}0{}.{}f".format(sign, max(minLength, 0), max(decimals, 0)) + "}"